+++ lib/commands/thor/service_module.rb
@@ -442,7 +442,7 @@ module DTK::Client
       # DTK-1768: removed below; and replaced by removing old dir if unequal to final dir
       # was not sure why clause namespace  was in so kept this condition
       #FileUtils.rm_rf(old_dir) unless namespace
-      if old_dir != module_final_dir
+      if old_dir and (old_dir != module_final_dir)
         FileUtils.rm_rf(old_dir) unless namespace
       end
       DTK::Client::OsUtil.print("Module '#{new_module_name}' has been created and module directory moved to #{repo_obj.repo_dir}",:yellow) unless namespace
